가지급금 양식 구매비 코드 추가 및 보고서 수정 item_no 관련 오작동 수정(수정 덮어쓰기)

+++ src/main/java/kr/letech/aprv/service/impl/AprvMngServiceImpl.java
@@ -471,8 +471,10 @@ public class AprvMngServiceImpl implements AprvMngService {
 
 			//기존 정보 삭제
 			for(String delNo : delArr) {
-				params.put("item_no", delNo);
-				docDAO.deleteTmpPayItem(params);
+				if(StringUtils.isNotEmpty(delNo)) {
+					params.put("item_no", delNo);
+					docDAO.deleteTmpPayItem(params);					
+				}
 			}
 			// 기존 정보 : 수정 / 신규 정보 : 추가
 			for (TmpPayItemVO tmpPayItemVO : tmpPayItemList) {
